+++ CHANGELOG.md
@@ -1,10 +1,22 @@
+## 0.0.6
+
+- Documentation extended, swapped the order of a couple of arguments in a method
+
+## 0.0.5
+
+- Changed `Env.expand()` to return just an expanded string rather than a tuple, added fully functioning example under `examples/`, fixed a bug for an empty EnvFilter check, fixed example in README.md
+
+## 0.0.4
+
+- A fix for importing the package.
+
 ## 0.0.3
 
 - Initial release to PyPI.
 
 ## 0.0.2
 
-- Fixed __main__.py by moving the most of the code under _def main()_
+- Fixed `__main__.py` by moving the most of the code under _def main()_
 
 ## 0.0.1